Sync up with latest changes

+++ app/controllers/users_controller.rb
@@ -1,5 +1,5 @@
 class UsersController < ApplicationController
-  before_action :set_user, only: [:show, :edit, :admdel, :memsdel]
+  before_action :set_user, only: [:show, :edit, :admdel, :memsdel, :update]
   wrap_parameters :user, include: [:username, :email, :password, :password_confirmation, :role]
 
   # GET /users
@@ -38,26 +38,31 @@ class UsersController < ApplicationController
   end
 
   def addrooms
-
   end
 
   def admview
     @user = User.new
-    @admall = User.find_by_role("admin")
   end
 
   def memsview
     @user = User.new
-    @memsall = User.find_by_role("user")
   end
 
-  def memshist
-    @user = User.new
-   # @memshistall = BookingHistory.find_by_role("username")
-  end
   def admincreation
     @user = User.new
   end
+
+  def adminupdate
+    @user = User.find_by_username(params[:user][:username])
+    respond_to do |format|
+      if @user.update_attribute(:role, 'admin') or @user.update_attribute(:role, 'user')
+        format.html { redirect_to dum_path, notice: 'Updated successfully'}
+      else
+          format.html { redirect_to admincreation_path }
+       end
+    end
+  end
+
   # GET /users/new
   def new
     @user = User.new
@@ -75,13 +80,15 @@ class UsersController < ApplicationController
     @user.role = 'user'
     respond_to do |format|
       if @user.save!
-        sign_in @user
-        flash[:notice] = "#{params[:user][:username]} User was successfully created."
-        format.html {redirect_to dumget_path}
+        flash[:notice] = "#{session[:username]} User was successfully created."
+        if session[:user_role] == 'admin'
+          format.html {redirect_to dum_path}
+          else
+            format.html {redirect_to root_path}
+            end
         #format.json { render :dum, status: :created, location: @user }
       else
-        flash[:notice] = "hi"
-        format.html { render :new }
+        format.html { render :root }
         #format.json { render json: @user.errors, status: :unprocessable_entity }
       end
     end
@@ -91,31 +98,53 @@ class UsersController < ApplicationController
     #@user_update = User.new(user_params)
     #@user = User.all.select {|u| u.username == params[:username]}
     #@user = User.find_by_username(params[:user][:username])
-    @user = User.new(user_params)
+    #@user.(user_params)
+    #@user.username = session[:user_name]
     #@user = User.find_by_username(params[:user][:username])
     respond_to do |format|
-      if @user.update(user_params)
+      if @currentuser.username != "Superadmin"
+        if @currentuser.update(user_params)
         flash[:success]="Profile updated"
-        format.html { redirect_to @user, notice: 'User was successfully updated.' }
+        if @currentuser.role == "admin"
+          format.html { redirect_to dum_url, notice: 'User was successfully updated.' }
+        else
+          format.html { redirect_to booking_histories_url, notice: 'User was successfully updated.' }
+        end
       else
         flash[:notice]="Profile couldn't be updated. Please try again"
-        format.html { render :edit }
+        format.html { redirect_to booking_histories_url }
         #format.json { render json: @user.errors, status: :unprocessable_entity }
+        end
+        else
+        format.html { redirect_to dum_url, notice: 'Superadmin cannot be updated.'}
+        end
       end
+  end
+  def checkexitinguser
+    @user = User.find(params[:id])
+    if (@user != null)
+      flash[:notice]="User does not exist. Please sign up"
+    else
+      flash[:notice]="Please enter password"
     end
   end
 
   def destroy
-    @userdel = User.find_by_username(params[:user][:username])
-      if @userdel.destroy
+      @userdel = User.find_by_username(params[:user][:username])
       respond_to do |format|
-        format.html { redirect_to dumget_path, notice: 'User was successfully deleted.' }
-        format.json { head :no_content }
-        end
+      if (@userdel.username != session[:user_name]) and (@userdel.username != "Superadmin")
+        if @userdel.destroy
+          format.html { redirect_to dum_url, notice: 'User was successfully deleted.' }
+          format.json { head :no_content }
+          end
+        else
+          format.html { redirect_to dum_url, notice: 'User cannot be deleted.' }
+      end
     end
-end
+  end
+
   def set_user
-      #@currentuser = User.find(session[:user_id])
+      @currentuser = User.find_by_username(session[:user_name])
   end
 
     # Never trust parameters from the scary internet, only allow the white list through.
